# This is a combination of 2 commits. # This is the 1st commit message:
feat: now cost claim is executed after wallet login. Now the form is reset when change the network

feat: now cost claim is executed after wallet login. Now the form is reset when change the network

# This is the commit message #2:

feat:  Added claim cost notice to the bottom side of the CrossForm component to display when the token is selected.